+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "spiral_sequence",
+  "version": "1.0.0",
+  "description": "Небольшой скрипт, генерируюший матрицу размерности 2*N - 1, заполненную случайными числами, и выводящий значения матицы из центра по спирали.",
+  "main": "lib.js",
+  "directories": {
+    "test": "test"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^5.2.0"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/AdilSerker/spiral_sequence.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/AdilSerker/spiral_sequence/issues"
+  },
+  "homepage": "https://github.com/AdilSerker/spiral_sequence#readme"
+}
